creation of log_simulator to write on a file with a certain pace during a certain time in a certain sequence

+++ source/log_writer.py
@@ -9,17 +9,53 @@ from random import randint
 import time
 import reader
 import os
+import datetime
+
+
+class LogSimulator(Thread):
+    def __init__(self, config_path):
+        Thread.__init__(self)
+        self.config_path = config_path
+
+    def run(self):
+        try:
+            config_file = io.open(self.config_path)
+        except Exception:
+            raise
+            # print("wrong config path '{}' for LogSimulator".format(self.config_path))
+
+        param_dict = {'erase_first': True}
+        for line in config_file:
+            if not line.strip().startswith('#'):
+                if "=" in line:
+                    key, value = (word.strip() for word in line.split('='))
+                    if key in ['log_path', 'line_type']:
+                        param_dict[key] = value
+                else:
+                    param_dict['pace'], param_dict['timeout'] = [int(i) for i in line.split(',')]
+                    print(param_dict)
+                    log_w = LogWriter(**param_dict)
+                    log_w.start()
+                    log_w.join()
+
+                    if param_dict['erase_first']:
+                        param_dict['erase_first'] = False
 
 
 class LogWriter(Thread):
-    def __init__(self, log_path, mode='HTTP', pace=1000, timeout=-1):
-        """Writes line in the file given by the log_path"""
+    """Writes a given type of lines in the file given by the log_path at a given speed during a given time"""
+    def __init__(self, log_path, line_type='HTTP_slow', pace=3000, timeout=-1, erase_first=True):
+        """
+        Args
+        ----
+        """
         Thread.__init__(self)
 
         self.log_path = log_path
-        self.mode = mode
+        self.line_type = line_type
         self.pace = pace
         self.timeout = timeout
+        self.erase_first = erase_first
 
         self.should_run = True
         self.log_file = None
@@ -27,19 +63,18 @@ class LogWriter(Thread):
     def run(self):
         line_count = 0
         start_time = time.time()
-        try:
-            os.remove(self.log_path)
-            print("Log file found and erased")
-        except IOError:
-            print("Log file not found, will be created")
-        self.log_file = io.open(self.log_path, 'wt')
+        if self.erase_first:
+            try:
+                os.remove(self.log_path)
+                print("Log file found and erased")
+            except Exception:
+                print("Log file not found, will be created")
+        self.log_file = io.open(self.log_path, 'at')
 
-        if self.mode == 'HTTP':
-            random_line = random_local_URL
-        else:
-            def random_line(): return 'line' + str(line_count)  # PEP8 says 'do not assign lambda expressions'...
-            # faster than ''.join(['line', str(line_count)])
+        # the function that will generate the lines is bound once for all
+        random_line = random_log_line(self.line_type)
 
+        # the expression used in the first while is defined once for all
         if self.timeout == -1:
             def timeout_check(): return True
         else:
@@ -48,12 +83,17 @@ class LogWriter(Thread):
         while self.should_run and timeout_check():
             line_count_second = 0
             start_time_second = time.time()
+            date = datetime.datetime.utcnow().strftime('[%d/%b/%Y:%X +0000]')  # ' {:+03d}00]'.format(randint(-12, 12))
             print('start :', start_time_second)
-            while line_count_second < self.pace:
-                self.log_file.write(random_line() + '\n')
+            while line_count_second < self.pace and self.should_run:
+                self.log_file.write(random_line(date=date, line_count=line_count) + '\n')
                 self.log_file.flush()
                 line_count += 1
                 line_count_second += 1
+                if line_count % 1000 == 0:
+                    date = datetime.datetime.utcnow().strftime('[%d/%b/%Y:%X +0000]')  # ' {:+03d}00]'.format(randint(-12, 12))
+                    # print("reset date at", line_count)
+
                 # print('written line', line_count-1)
                 # time.sleep(1.0/self.pace)
             delta = time.time() - start_time_second
@@ -61,17 +101,19 @@ class LogWriter(Thread):
             if delta < 1:
                 print('waiting for :', 1-delta)
                 time.sleep(1-delta)
+            else:
+                print('writing operations are too slow, change line_type or reduce pace')
             print('end :', time.time(), 'line_count :', line_count, '\n==========')
 
         print('LogWriter end after {} lines in {:.4f}s'.format(line_count, time.time()-start_time))
         self.log_file.close()
 
 
-def random_local_URL(factor=3, max_depth=4):
+def random_local_URL(factor, max_depth):
     """Return a random local URL
 
-    NB number of section = len(section) * factor, by default = 9
-    NB max_depth is the highest number possible of /section/subsection/subsubsection/etc., by default = 4 (and min = 0)
+    NB number of section = len(section) * factor
+    NB max_depth is the highest number possible of /section/subsection/subsubsection/etc.
     Ex /archive2/blog0/archive2/page1/ -> depth = 4
 
     Examples
@@ -82,8 +124,9 @@ def random_local_URL(factor=3, max_depth=4):
     /archive0/blog1/
     /blog0/picture.png
     /page0/archive1/archive2/index.html
-    /archive2/archive0/archive2/archive2/form.php    <- max_depth with URL_end
+    /archive2/archive0/archive2/archive2/form.php    <- max_depth = 4, with URL_end
      """
+    # print(factor, max_depth)
     section = ['page', 'blog', 'archive']
     URL_begging = [word + str(i) for i in xrange(factor) for word in section]
     URL_end = ['picture.png', 'index.html', 'form.php']
@@ -91,16 +134,88 @@ def random_local_URL(factor=3, max_depth=4):
                           [URL_end[randint(0, len(URL_end)-1)]*(randint(0, 4) >= 1)])
 
 
+def random_HTTP_request(**kwargs):
+    method = ['HEAD', 'GET', 'OPTIONS', 'TRACE', 'POST', 'PUT', 'DELETE', 'PATCH', 'CONNECT'][randint(0, 8)]
+    return '"{} {} {}"'.format(method, random_local_URL(**kwargs), 'HTTP/1.1')
+
+
+def random_log_line(line_type, **kwargs):
+    if line_type == 'HTTP_fast':
+        print("generating fast")
+        request_param = {'factor': 4, 'max_depth': 2}
+        request_param.update(kwargs)
+
+        def fast_rand(date=0, line_count=0):
+            remote_host = "123.12.45.78"
+            remote_log_name = '-'
+            auth_user = '-'
+            request = random_HTTP_request(**request_param)
+            status = '100'
+            bytes_ = '1500'
+            return " ".join([remote_host, remote_log_name, auth_user, date, request, status, bytes_])
+        return fast_rand
+
+    elif line_type == 'HTTP_slow':
+        print("generating slow")
+        request_param = {'factor': 4, 'max_depth': 4}
+        request_param.update(kwargs)
+
+        def slow_rand(date=0, line_count=0):
+            remote_host = ".".join([str(randint(0, 255)) for i in xrange(4)])
+            remote_log_name = '-'
+            auth_user = '-'
+            request = random_HTTP_request(**request_param)
+            status = str(100*randint(1, 5) + randint(0, 5))
+            bytes_ = str(randint(0, 10e4))
+            return " ".join([remote_host, remote_log_name, auth_user, date, request, status, bytes_])
+        return slow_rand
+
+    elif line_type == 'line':
+        print("generating line")
+
+        def line_rand(date=0, line_count=0):
+            return 'line' + str(line_count)  # PEP8 says 'do not assign lambda expressions'...
+            # faster than ''.join(['line', str(line_count)])
+        return line_rand
+
+    else:
+        raise ValueError('line_type={} is an incorrect parameter'.format(line_type))
+
 if __name__ == '__main__':
     # import reader
-    log_path = '../data/wtest'
-    lw = LogWriter(log_path, mode='HTTP', timeout=1, pace=5000)
-    rd = reader.ReaderThread(log_path)
-    lw.start()
-    time.sleep(0.1)
-    rd.start()
-    time.sleep(5)
+    # #log_path = os.path.expanduser('~') + '\log.txt'
+    # log_path = '../data/wtest'
+    # pace = 5
+    #
+    # lw = LogWriter(log_path, line_type='line', timeout=1, pace=pace)
+    # lw.start()
+    # lw.join()
+    #
+    # lw = LogWriter(log_path, line_type='HTTP_fast', timeout=1, pace=pace)
+    # lw.start()
+    # lw.join()
+    #
+    # lw = LogWriter(log_path, line_type='HTTP_slow', timeout=1, pace=pace)
+    # lw.start()
+    # lw.join()
+
+    # f = io.open('../data/sim_config', 'wt')
+    # f.write('# write log lines at the speed \'x\' (per seconds) during \'t\' seconds\n')
+    # f.write('10, 3\n')
+    # f.write('50, 2\n')
+    # f.flush()
+    # f.close()
+
+    ls = LogSimulator('../data/sim_config')
+    ls.start()
+
+    # rd = reader.ReaderThread(log_path)
+    # time.sleep(0.1)
+    # rd.start()
+    # time.sleep(5)
     # rd.should_run = False
+
+    # print(random_log_line())
 
     # import timeit
     # print(timeit.timeit("''.join(['line', str(1234)])", number=1000))
